Support ST_TIME_NS semantic type w/ INT64 as an alternative to TIME64NS in the UI.
Summary: Part of a series of diffs that will be used to switch over to ST_TIME_NS isntead of using TIME64NS. The plan here was to support both the native data type and the semantic type simultaneously. In the future, we will deprecate the native data type.

Test Plan: Tested w/ a backend where time64ns is already switched over to INT64 + ST_TIME_NS and this wokred like a charm.

Reviewers: nick, vihang, michelle, nserrino, jamesbartlett, #engineering, zasgar

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D6706

GitOrigin-RevId: 01f74ebb2f608c65ddc0f6954d17c67c679a5190