Fix cancel state function setter.
Summary:
state setters allow you to use functions to set the state so you can use the current state to make the next state. When your state variable is a function, you must nest the function in another function

This is a missing fix I guess. Before this, couldnt' get any results on any scripts.

Test Plan: tested on prod and works.

Reviewers: nick, vihang, michelle, #engineering

Reviewed By: nick, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6708

GitOrigin-RevId: 8852fce748943e09b697003e8f724d80da11b4be