Behave like `{head,tail} -f` if user scrolls to the edges of a data table.
Summary:
When new data enters a data table, if that table was scrolled to the top or bottom of its range, it sticks to wherever that edge moved. This is useful when a streaming query outputs an ever-flowing data set like HTTP requests and the user wants to watch that output live.

When scrolling to the top or bottom of a data table, stick to that edge if new data comes in.

This is especially useful for streaming queries, where the user may wish to monitor a live stream of recent events. Think `tail -f`. A future change will also put a finite scroll buffer on this for performance.
This is a first iteration. Depending on how this experience feels to users, we may try adding UI elements to enable/disable this mode directly or offer other methods.

Test Plan:
Try this streaming query. Sort it such that new data comes in either at the top or at the bottom, scroll to that edge, and release. The table should pin the scrollbar to that edge until you move it away.
```
import px
df = px.DataFrame(table='http_events', start_time='-2s')
df.pod = df.ctx['pod']
df = df.drop(['time_', 'upid', 'trace_role', 'http_minor_version', 'http_content_type', 'http_resp_message'])
px.display(df.stream())
```

Reviewers: nserrino, michelle, #engineering, vihang

Reviewed By: #engineering, vihang

Subscribers: philkuz, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6713

GitOrigin-RevId: 2cc67d26cf1df7aa467937a63d3a6a7aa0e010d1