Propagate request building errors in the observable model of the Vizier GRPC client
Summary: Request building can create errors that were not properly handled. I moved the request builder in the observable function callback, but need some feedback on why we choose to send stuff around. A reviewer should help with that.

Test Plan: Tested with an erroneous vis spec where we had an integer instead of a string in the defaultValue and the error showed up, whereas the error would be emitted otherwise.

Reviewers: nick, vihang, michelle, #engineering

Reviewed By: nick, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6716

GitOrigin-RevId: bf79ab25eb0a164482f95f63b42a8af56fd544e4