Change px.sum of booleans to return int
Summary:
This diff changes the behavior of px.sum on booleans to return a sum rather than a logical OR.

We need a consistent definition of what the sum of two booleans means. So I'm leaving this as a preview until we can make that call.

If we decide boolean + boolean = boolean, then this diff is not correct. Note that different languages (or even libraries within languages) have differing approaches.

The right long-term solution is to do this via data type casts. So this could be a temporary workaround until that is supported.

Test Plan: Modified the tests

Reviewers: philkuz, nserrino, #engineering, zasgar

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D6728

GitOrigin-RevId: 1b0c3a321276b1fb6871991d16d0421edbbec9f9