Don't allow cancelling a query that is already cancelled
Summary: As every query run and retry cancels the previous one, the user cancelling one script then trying to run another that has mutations was triggering a second cancel on the first query. Due to awkward timing, that was setting loaded state to false a second time right after the second query started, so the tracepoints dialog thought it needed to hide. The fix is for the cancellation function to run once and replace itself with a noop.

Test Plan: Run `px/cluster` with a long start_time, like an hour. Cancel it before it returns. Then try to run `tcp_retransmits` with fresh tracepoints. The visual loading indicator should stick around for the correct duration.

Reviewers: nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6727

GitOrigin-RevId: e0e83aef3549c02676d6c462c5f5948c5ba23560