PP-2310: Add Carnot UDF to create a script reference type for custom deep linking in Live Views
Summary:
For our DNS script and other use cases, we want to be able to add custom dynamic deep links within a script. Currently our deep links are exclusive to entity types like pod, but we want to add a capability to PxL for the user to specify a deep link in an arbitrary output column.

This diff adds the 'px._script_reference' UDF which will be responsible for taking a script name, a text label, and script args and returning a stringified JSON value like this:

```
{
	label: 'visible link name',
	scriptID: 'px/foo',
	args: {
		start_time: '-5s',
		entity_arg: 'val_from_col'
	}
}
```

This is a bit challenging due to our lack of support for object types in PxL, since creating the above, especially when the script arg comes from a colum in a DataFrame, is a natural fit for an object type in PxL. Once that support comes in, this function will need to be refactored, but I am trying to minimize the amount of unnecessary work while still supporting the requested use case.

In addition to a lack of object types in PxL, we also have a lack of support for variadic UDFS. A function like https://docs.snowflake.com/en/sql-reference/functions/object_construct.html is very difficult to support in our UDF catalog today, so that will be another thing that needs to be refactored once we support object types. Due to this limitation, this diff adds support for up to 4 script args (we don't use that many in most scripts today), but when variadic UDF args are supported in our catalog, we will update this function to use them and remove that limitation.

In the next few diffs, I will

1. Add a semantic type for Script reference (ST_SCRIPT_REFERENCE) and ensure the output of this function is that semantic type.

2. Add handling in the compiler for a public API for this function, which will probably look like

```
df.col = px.ScriptReference(label='visible link name', script='px/foo', args={
	'start_time': '-5s',
	'entity_arg': df.foo
})
```

3. Add support in the UI to hyperlink columns which are the ST_SCRIPT_REFERENCE semantic type.

Test Plan: added unit test

Reviewers: philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

Subscribers: htroisi

JIRA Issues: PP-2310

Differential Revision: https://phab.corp.pixielabs.ai/D6753

GitOrigin-RevId: 00a3c541ac9eaf92466486024f01d87b98e3c21b