PP-2310: Add semantic type for script reference
Summary: Depends on D6753. This diff adds ST_SCRIPT_REFERENCE as a semantic type so that the UI can interpret columns of this type and render them as deep links. It also adds the semantic type to the UDF added in the prior diff.

Test Plan: run in skaffold

Reviewers: philkuz, jamesbartlett, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2310

Differential Revision: https://phab.corp.pixielabs.ai/D6759

GitOrigin-RevId: d965486c484b51ae5cf553d160d8b9fc06f2ccdd