Add StitchFrames() function template as the common interface for stitching requests & responses of all protocols
Summary:
Also rename ProcessFrames() -> StitchFrames().

This matches the ParseFrame() pattern.

The benefit is that it now is easier to understand the expectation on the ProcessFrames()
API. Previously, ProcessFrames() is only linked, it was obscure to learn how it's invoked.

In a followup diff, the common interface templates will be moved into one single header file.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6763

GitOrigin-RevId: 2a710bc4c9368645b05eeea2bb7e7d218e5f216d