Less scary messaging for auth errors
Summary:
Displays the Pixienaut being chased by an octopus instead of a big red triangle + exclamation point. Also adds an option to render with the Pixienaut on the toilet, in an "oh no" pose. Messaging for sign in, log in, and general auth errors made more friendly and given a call to action (navigating) button.

The toilet pose is not in use yet. The goal going forward is to use the octopus for recoverable errors and the toilet for fatal errors.

The octopus box:
{F104225}

The toilet box (not in use at this time, but fatal errors are welcome to use it going forward):
{F104226}

Test Plan: To trigger the log in error, try to log in with a Google account that isn't registered with us. To trigger the sign up error, try to register an account that is already registered.

Reviewers: nserrino, michelle, #engineering, vihang

Reviewed By: #engineering, vihang

Subscribers: vihang

JIRA Issues: PC-575

Differential Revision: https://phab.corp.pixielabs.ai/D6769

GitOrigin-RevId: 8c4e9cd3163f10e53a3ebafd82dd389ed0930cdc