PP-2310: Add the public PxL API for creating script reference deep links
Summary: Depends on D6759. This diff adds the public PxL function for creating a custom deep link in a table (as what we are now calling a script reference). I also decided to bump up the max number of supported script arguments in a script reference from 4 to 8 based on the feedback that we would be using 4 right out of the gate. Next up, I will add the ST_SCRIPT_REFERENCE-specific rendering to the UI so that these script references show up as clickable deep links in the UI.

Test Plan: added unit tests

Reviewers: philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2310

Differential Revision: https://phab.corp.pixielabs.ai/D6771

GitOrigin-RevId: 1df6eee4f5f7d70307e2e5e1f0de54cd519de72c