Fix incorrect typing (which compiled anyway) in two files
Summary: For `callback.tsx`, the issue was that React Router's `<Link>` component supports forwarded props, but does not specify that fact in the typings. Fixed by using a trivial HOC. In `convert-to-vega-spec.ts`, we are using a prop that does not exist in non-patched Vega, but which does in our patched version. Adding a type extension to define it fixes this. In both scenarios, the issue was only with the typings and the compilation succeeded anyway as the code is fine. Also, removes an unused interface.

Test Plan: Run the dev build.

Reviewers: #engineering, nserrino, michelle, vihang

Reviewed By: #engineering, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6785

GitOrigin-RevId: 94ffca0068a6b39104aedff7ad8e3bd592f14c23