Add UserSettings table
Summary:
we want to add user settings so that we can track whether or not a user has seen the onboarding tour overlay.
this will also be used for other settings in the future, such as hotkeys.

this diff only adds the table to the profile service, but doesnt use it yet.
the following diffs will:
- add userSettings to the grpc API
- add userSettings to the gql API
- implement the API in the backend

followed the typical userSettings design, where the key/value are all strings.
it is up to the consumer (UI in this case) to know the type of the userSetting it has just requested and do typecasting if needed.

Test Plan: ran plc-dev and made sure the migration goes through

Reviewers: nick, zasgar, nserrino, #engineering, vihang

Reviewed By: #engineering, vihang

Subscribers: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6766

GitOrigin-RevId: cec9c18e9df27827f1c21714ea161ecb6704de14