Fix UI flags for nginx deploys
Summary:
the previous change broke the UI for ngninx, because WEBPACK_REPLACE was an undefined variable.
there were a few solutions that immediately came to mind:
- update webpack non-dev server builds to set WEBPACK_REPLACE to false. didn't like this, since we didn't want to use webpack replace/define plugin for non-dev server builds in the first place.
- update nginx to set WEBPACK_REPLACE to false. also didnt like this, because now nginx has to know about what should be a webpack specific variable.
we originally wanted the replaced values to be strings so that the code would still build even if there was no replace. but with the webpackDefinePlugin, this looks like we can't do this anymore. so, we might as well make them all variables rather than strings.

Test Plan: deployed to staging and ran yarn dev. verified that both builds are functional

Reviewers: vihang, zasgar, nick, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6795

GitOrigin-RevId: 799aa0a0536b5d4fd173decf6416d01ac2a9b8c2