Surface agent pod names
Summary:
we updated our logging scheme so that we can view logs for pods, given the cluster ID and pod name.
however, we previously didnt surface up the pod name anywhere, which makes it impossible for us to look at PEM logs.

Test Plan: unit tests, ran vizier with skaffold and verified it shows up correctly in the script

Reviewers: zasgar, nserrino, philkuz, #engineering, vihang

Reviewed By: #engineering, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6800

GitOrigin-RevId: 92c631131f76c00a513c7f72111ed933bf169dea