Add gql schema for fetching/modifying UserSettings
Summary:
we are adding user settings to our system. the previous diff added the userSettings table that we will use to store userSettings in the database.
this diff adds the gql schema (currently unimplemented) which the UI will use in order to get the user settings.
to get some userSettings, you can pass in multiple keys which you want to fetch.
to update userSettings, you can pass an array of keys, and array of values that you want to update those keys to. the backend will check that the length of these arrays are equal.

Test Plan: unit tests pass. there is no actual implementation yet.

Reviewers: nick, nserrino, zasgar, vihang, #engineering

Reviewed By: nick, nserrino, vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6809

GitOrigin-RevId: f81636cd292ea0ffbaa124bf74ab4d0b27aa7de9