Adding a fix to substring that makes sure that we dont' error out in the execution
Summary: Substring previously failed when you hit these edge cases that I showed in the test, meaning you'd get an execution error which sucks for a user. This fixes that problem by just returning an empty string, which might be confusing but at least is better than failing. An alternative is to actually encode an error in the returned string, but not completely sure if that's what we want to do eiher.

Test Plan: Added tests for the edge cases.

Reviewers: nserrino, zasgar, michelle, vihang, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6810

GitOrigin-RevId: ec54ddc186b7761ae56fc99b205a1f55ea96daa9