Underline deep links for greater discoverability
Summary: In multiple customer conversations, the user was unaware that entity links could be clicked in order to navigate to that entity in the Live View. While it may look a little less nice, it seems like making deep links resemble traditional links more could improve the discoverability of that feature.

Test Plan: n/a

Reviewers: zasgar, michelle, nick, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6812

GitOrigin-RevId: bf2853cf17414a3742bede41b5f5e1d29907cd4e