Get rid of HTMLWebpackHarddiskPlugin
Summary:
This plugin was writing the `index.html` asset to disk before
the asset was ready to be emmited. That combined with the new `compareBeforeEmit`
option was causing the asset to not be emitted and hence not be included
in the archive.
Removing the plugin fixes the issue. (AFAIK the plugin doesn't do anything useful
in non-devserver mode and I don't believe we rely on an index.html in devserver mode)

While we are at it, set `compareBeforeEmit` to false in `prod` builds.
Prod builds are done in a "clean room" anyway so this shouldn't matter but
this will help the poor future engineer who's trying to quickly fix some issue
for a release, and happens to have not cleaned out their `dist` folder before
attempting a build.

Test Plan: `rm -rf dist/*` followed by `make build_{dev,prod}`

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6808

GitOrigin-RevId: 18c2c8c7b299f472473b68371a806721fe148ddf