Stirling: Extract correct kernel version on Debian
Summary:
Stirling does not deploy properly on Debian installations without Linux headers. This is because the Linux version that is detected is wrong,
and this results in the wrong prepacked headers config. Ultimately that causes BPF probes to fail to deploy.

Test Plan: Test cases added using Debian /proc system. Also manually tested on a Debian VM.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6813

GitOrigin-RevId: dc2c764cdfe5b771fa54e2a272cf84d740f0a4e4