Don't leak expired query cancellation functions.
Summary: This was causing the tracepoint mutation modal to dismiss itself prematurely. The live view's loading state is marked false when a query completes or is cancelled, and the first step of running a query is cancelling the already-running one. It now won't try to do this if there is nothing to cancel.

Test Plan: Run a script with no mutations to completion, like `px/cluster`. Then run one with mutations like `px/tcp_drops`. The tracepoints modal should no longer dismiss itself prematurely. The previous fix for this only fixed it when the first script was cancelled, but not if it completed normally.

Reviewers: nserrino, michelle, vihang, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6836

GitOrigin-RevId: 1d6ddfec178ef100d39d97ab9931b3fea32198e3