Correct semantic errors in local GQL modifications; add graphclconfig file.
Summary: the GraphQL language service (used by plugins for editors and IDEs) marks the entire schema invalid if it detects syntax or semantic errors, making the tool think every query is invalid/unknown. Fixing semantic errors takes care of that. Additionally, we have two schemas with some overlap, which were both being seen by that tooling. Using  `.graphqlconfig` file to ignore the one that is a subset of the other takes care of this as well. Both fixes combined allow tooling to recognize valid queries. Notably, this does not change the application's behavior at all: it just makes sure editors and IDEs understand what's happening.

Test Plan: Check out the various usages of the `gql` template tag in our TypeScript code, in an editor/IDE that can highlight and inspect the syntax. It should now succeed when trying to do so. The build should be entirely unaffected.

Reviewers: nserrino, michelle, vihang, #engineering, zasgar

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6838

GitOrigin-RevId: 5fe4cf6418c5e834f26c2be8f5621df12520ef93