ConnectorContext optimizations
Summary:
Store some information when the connector is created, so that repeated calls to GetUPIDs() doesn't do a bunch of extra work.

In stirling_wrapper with a StandaloneContext, this reduces the perf from ~75% CPU to ~20% CPU usage.

Likely doesn't affect stirling the AgentContext (i.e. in the PEM) in any significant way,
since the AgentMetadataState was already snapshotted in AgentContext. So only expect marginal improvements there.

Test Plan: Existing tests.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6841

GitOrigin-RevId: 4d5f30fc86a76300f44f2eadd9fe82024767cff0