Add yaml files for sourcegraph svc/ingress
Summary:
these are the files needed to set up the ingress/certs for sourcegraph.
deploying source graph itself requires A LOT of yamls, which are all located in the deploy-sourcegraph repo. i didnt think it was necessary to duplicate those here.

Test Plan: deployed, got a working  version of sourcegraph.

Reviewers: zasgar, vihang, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6848

GitOrigin-RevId: f4cbfa27e073ecc7d232935d6f312c86d6885c8a