Update profile service to handle Get/UpdateUserSettings
Summary:
we are adding user settings in our backend, so that we can use usersettings to determine whether or not to show the onboarding overlay.
this part updates the profile service to handle getting/updating user settings.
this diff:
- updates the profile proto with the new rpc methods
- updates the profile server to handle the new rpc calls
- updates profile datastore to actually get/update the values in the db

the last part in the backend is to update the API service to make this call to the profile service.

Test Plan: unit tests

Reviewers: vihang, nserrino, nick, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6849

GitOrigin-RevId: 649b956e68125348c3cd6e32ce50d65d4d2c40f5