PP-2318: collect statistics about inbound k8s updates and agent heartbeats for metadata service.
Summary:
This diff adds a stats handler to the metadata service which watches the k8s updates and agent heartbeats and notes the volume received by each.
It breaks down the k8s updates by entity type (pod, etc) and the agent heartbeats by agent data info, created processes, and terminated processes.
We are adding this in order to better understand the load on the metadata service for a series of stability improvements that we want to make.

Also for some reason make gazelle changed an autocomplete BUILD.bazel ordering, so added that too.

Test Plan: ran it in my cluster and verified the correct information was printed out

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2318

Differential Revision: https://phab.corp.pixielabs.ai/D6859

GitOrigin-RevId: 04e3edbc8f4855d7a3b4045b0caa33394978fc41