Periodically prune computedSchema of any deleted agents
Summary:
we are seeing an issue occuring on clusters (especially larger clusters), where the computedSchema grows too large and can no longer be sent to etcd.
after some examination, it appears we are hitting a race condition where if an agent is deleted the same time as another is updating its schema, its possible for zombie agents/tables to be left behind in the computed schema.
we could add locks, but that would introduce a performance impact since the lock would have to be around a network call to etcd.
we are also heavily refactoring the metadata service, so most of this logic will be removed in the future anyway. we decided that a temporary fix, where we periodically clean up the computedSchema, should suffice until our refactor

Test Plan: unit test, ran Vizier with skaffold and verified the zombie agent IDs and tables were pruned. also verified that I could still run scripts afterwards

Reviewers: nserrino, zasgar, philkuz, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6860

GitOrigin-RevId: b34109581979dd58d295f6b4bc57f1cf45987253