Fix cancel button for tracepoint deployment
Summary: Not the first bug that came from having to make a fix in both the mutation and normal path. I'll be putting some thought into a more maintainable structure for this file; the behavior is already what we want.

Test Plan: Try to cancel tracepoints being deployed. `tcp_drops` is a good example, and renaming the UpsertTracepoint parameter is sufficient to force it to deploy new tracepoints/see the dialog again.

Reviewers: nserrino, michelle, vihang, #engineering

Reviewed By: nserrino, michelle, #engineering

JIRA Issues: PC-694

Differential Revision: https://phab.corp.pixielabs.ai/D6890

GitOrigin-RevId: 9bbc53fcd4c6da47529d350e75b6a5571c6b7a66