Automatically open the tour modal the first time a user interacts with Pixie through Vizier.
Summary: This waits for a user setting to load, and if it sees that the user has never seen the tour, it marks that flag and shows them the tour. It will not do this again unless something clears that setting - you'll need to use a debugger to reset it for now. In the future when we offer a settings dialog for other preferences, a button can exist there to reset any tours/tips/etc.

Test Plan: Open the frontend and note that the tour modal appears fairly quickly by itself - but only ever does this once.

Reviewers: nserrino, michelle, vihang, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-651

Differential Revision: https://phab.corp.pixielabs.ai/D6898

GitOrigin-RevId: 4dfe877a0c44d870563e5c0c80df7442672bb4d9