golang log filename and linenumber
Summary:
Configure logrus to also log the filename and line number of the log
message. This makes it easier to track down errors.
We need to update logrus to get this flag to report the actual file (otherwise
it reports the logrus/entry.go) on go `v1.14+`
The prefixed formatter also hasn't been updated in 3 years and doesn't support
logging filename and linenumber so just remove it. `pixie-cloud` logs still look sane.

This has a perf penalty but hopefully it's not bad enough to be a problem for us.

Test Plan: Deploy pixie cloud and look at logs for various services.

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6903

GitOrigin-RevId: d9660b684270ea1781f5bd988e92be771119c1dc