Cleanup PV usage
Summary:
This fixes two problems:
1) `postgres` used to use a manually created PV. This meant that the reclaim
policy for the PV was set to `Retain` and once the PVC was terminated, the PV
had to be manually cleaned up before another cloud deploy would work. Instead
just go back to letting the PVC autocreate the PV. Note: We need to set a `subPath`
in the volume mount because postgres doesn't like using a mount path as the data dir.

2) `elastic` used to provision `ssd` storage even when running in dev mode. Since
the `ssd` storage class provisioner was set to GKE, running `cloud` on `minikube`
wouldn't work. This clears out the storageClass and uses the patch files to
set it to `ssd` in `prod` and `staging` and to `standard` in `dev` mode. This should
help unblock deploying cloud on minikube for dev test environments.

Test Plan: Deploy cloud deps. Load dev db. Skaffold deploy cloud.

Reviewers: michelle, zasgar, jamesbartlett, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6911

GitOrigin-RevId: fa3c82c1cbfb1014003940a7e7aeb9e46d722397