PC-702: px/cluster should be the landing page for a new user.
Summary: When users previously logged in for the first time, they would see a blaak page and the command input open. This was not a good user experience, we want to replace it with the px/cluster view. Due to some legacy code around how clusters get loaded, this change had to be a bit surgical and put in script-loader, which runs at the right time for this interception (it can't be a default value in script-context because the cluster is not known at that time). We should clean up this cluster logic since it causes some weird orders of events and structuring in the script/etc logic.

Test Plan: ran it with application cache fully cleared

Reviewers: michelle, nick, philkuz, #engineering, zasgar

Reviewed By: michelle, #engineering, zasgar

JIRA Issues: PC-702

Differential Revision: https://phab.corp.pixielabs.ai/D6918

GitOrigin-RevId: a874606bb9148c992e1b34e60aa42ee84c2356f8