Add *.pl-nats to DNS Names
Summary:
Once we converted NATS/STAN to a stateful set, the replicas peer
each other by connecting to `<pod-name>.pl-nats`. Since TLS is enabled on NATS
we need `*.pl-nats` to be in the list of DNS Names for the SSL Cert.

Test Plan: Run the StatefulSet version of NATS/STAN

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6923

GitOrigin-RevId: fc78d5b00adb2617d30d75a597d025537d7e718f