Jenkins job trigger fixes
Summary:
Tests did not run properly in https://phab.corp.pixielabs.ai/D6919?id=24800.

While the bazel query producing `bazel_tests_bpf`  was properly identifying the test to run, it was being gated by `bazel_cc_changed` which didn't realize that a shell test was involved.

As a result, made a few edits to decouple the bazel queries for the test targets from the cc check. Relatedly, also made the rule for the creation of docs more explicit, and decoupled from clang-tidy.

Not sure if everything here is 100% right, so please let me know if there was some intention behind some of the previous structure.

Test Plan: Jenkins/manual

Reviewers: #engineering, zasgar, vihang

Reviewed By: #engineering, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6922

GitOrigin-RevId: a170b1f18ff47e88f3369474bc36c5b443cd626a