PP-2331 Remove hostport from cloudconn deployment yaml
Summary:
We are seeing a problem in single-node clusters where the cloudconnector does not get deleted in an update/deploy. as a result, two cloudconnectors end up sticking around, with one in a pending state.
this is because the steps during an update/deploy are:
- Delete old vz resources except cloudconn.
- Deploy new vz resources, including cloudconn.
Usually what happens in this last step is that the cloudconn deployment is updated, so it launches a new pod with the new version. Once that pod starts up, k8s deployments automatically handle deleting the old pod.
The problem in the one-node cluster is that the new pod can't deploy because the port is already in use.
Taking a look at the deployment yaml itself, this is because we've specified the hostport and two containers using the same hostPort cannot be scheduled on the same node
we dont actually need hostport to be specified.

Test Plan: created an rc. tried deploys and updates on minikube + gke

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6933

GitOrigin-RevId: c440123f3643d6a6abb8e0d5a3ae5360fc777f0f