Use a better, more visible command hint and make it clickable
Summary: This uses a new SVG that shows the same in all browsers, makes it fully visible, a bit bigger, and turns it into a button. Clicking it does the same thing as pressing the keyboard shortcut that it's hinting at.

Test Plan: Open the command input, attempt to type any command, and try both Shift+Enter and the hint/button. In both cases, nothing should happen when there is no valid command entered, and the command will run if it is valid.

Reviewers: nserrino, michelle, vihang, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-705

Differential Revision: https://phab.corp.pixielabs.ai/D6932

GitOrigin-RevId: 8c74970d0ed8a994ef0e608fd3fdb5f7fb2f07a8