Add cluster heartbeat events to segment
Summary:
we want to track cluster heartbeats, so that we can get info about the clusters and their # of nodes without having to probe the db.
this currently sends an event for every heartbeat (5s) and is tracking it from the cloud-side since we don't have a segment source for Viziers. this does, however, make sending the event after X heartbeats more difficult, since we would have to track that info in vzmgr per Vizier. I think we should just try out the current method for now, and if the # of events is too overwhelming, we can add that in if need-be.

Test Plan: deploy to staging and try it out

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6936

GitOrigin-RevId: fbc9b8736c2bdd92d971c941c4b0249897e6ca5e