Add API key db schema to auth service
Summary:
we need api keys that can be used with the Pixie API.
in a previous diff, i originally added the API keys table to vzmgr, since that is where our deployment keys are.
however, after some clarification and further thinking, i think it makes more sense to be in the auth service.

Test Plan: n/a

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6943

GitOrigin-RevId: 060b68db5a4a4845fe9a1b650e50eadbd43e8a9f