Hide the only non-keyboard way to open the Ctrl+K command input.
Summary: This is to ensure it isn't discoverable anymore. The experience needs a lot of work, and we've chosen to hide it for now rather than introduce new users to it. Existing users can still press Ctrl/Cmd+K to open it.

Test Plan: In the live views, the PX logo in the breadcrumbs bar is no longer a button.

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-720

Differential Revision: https://phab.corp.pixielabs.ai/D6946

GitOrigin-RevId: 47fdacb26d7f8caf35a323f051ad067c78c5ec84