Update cluster token issued at time
Summary:
we were seeing an issue on a lot of customer clusters where they would occasionally hit a "stream read error". upon closer examination, this would turn out to be a "invalid auth token: use before issue" error.
this is because on passthrough mode, we generate cluster tokens on the cloud side. these tokens are sent in the passthrough request and used as the token when executing healthcheck/executeScript queries.
the problem occurs when the time on the user's cluster is earlier than the cloud's time.
we should account for this by sending a token with an earlier "issued at".
this will not fix bugs where the user's cluster's clock skew is much larger... but that is not a problem we should aim to solve

Test Plan: deployed to staging, verified that my cluster can still run scripts.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6938

GitOrigin-RevId: 3951e671a631b46e86e1e0afa1f849867554b823