Fatal when STAN connection is lost
Summary:
This adds a connection lost handler to STAN, that fatals the caller
when then STAN connection is lost. This will cause the binary to terminate
and then reconnect to STAN when it is run again.

This is a worse hack than actually capturing the lost connection and trying
to reconnect, however since the connection is created in the servers but used by
the bridge after a few layers of redirection, that will require a bit more plumbing
to get working.

Test Plan:
Terminate the stan pods on a cloud deployment, watch `vzconn` fatal,
restart and observer that `cloud-connector` eventually ends up healthy wihtout
any manual intervention.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-1701

Differential Revision: https://phab.corp.pixielabs.ai/D6953

GitOrigin-RevId: cf3a2033b6bcef79d721b5936b062ce8a9550101