Update clusters resolver to make single call to db
Summary:
when fetching info about an org's viziers, we would make a call to vzmgr per vizier, asynchronously.
this slowed down the clusters request in the pixie org, since we were making 188 API + db calls in a single request.
there was a leftover TODO for parallelizing this  code... i figured making 188+ API calls to vzmgr was still suboptimal, when we should be able to make a single call to the db to fetch all of the data we need.
added a GetVizierInfos method for fetching the vizier info for multiple viziers. if requesting an invalid vizier, or a vizier not belonging to the org, we just return a nil VizierInfo. this is to avoid leaking info about existing vizierIds.

Test Plan: unit tests

Reviewers: zasgar, nserrino, #engineering, vihang

Reviewed By: #engineering, vihang

Subscribers: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6964

GitOrigin-RevId: 6606c1970f74d053accd479256a43410b8c8d8eb