Fix data-formatting in bar charts
Summary: in #ui @nserrino pointed out our bar charts would yield NaNUndefined for labels. This happened when a value was a metadata because the code was incorrectly trying to format the label. Simplified the interface to the internal bar chart implementation so that users directly manage the formatting function on the axis they want instead of the implicit / bug-prone solution before.

Test Plan: Tested with px/service_stats and natalie's script from #ui and both work.

Reviewers: nserrino, nick, michelle, vihang, #engineering

Reviewed By: vihang, #engineering

Subscribers: nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D6978

GitOrigin-RevId: b2e7dbcf43ae2d1386f068dd40b48bd96e0810ba