Resolve the evil demon soft lock bug.
Summary: More accurately, finish implementing an ugly hack that was placed in this loading code to fix, well, the same problem in different conditions.

Test Plan: Difficult to repro, but the bug was most common to trigger with a custom script being run from the editor, in Chrome.

Reviewers: nserrino, michelle, philkuz, #engineering

Reviewed By: nserrino, #engineering

JIRA Issues: PC-688

Differential Revision: https://phab.corp.pixielabs.ai/D6982

GitOrigin-RevId: fd2f519d1fa2f2960c4e902b73303c17ba3a4a81