Allow for a blank scratch pad script in the selection.
Summary:
As this isn't a real script upstream, there are quite a few special cases added for it.

{F107513}
{F107514}

Test Plan: In the live view, go to select a script. There should be a new item at the top of the list, that will not pre-select unless it's the only thing left from filtering. Choosing it will switch to a new script with some explaining comments. It should be just as ephemeral as edits to other scripts are.

Reviewers: michelle, #engineering, nick

Reviewed By: michelle, #engineering

Subscribers: htroisi, zasgar

JIRA Issues: PC-723

Differential Revision: https://phab.corp.pixielabs.ai/D6981

GitOrigin-RevId: 1254832b1bf06bd361660f00d53f6e0727f7f5dd