Accept API keys in request pixie-api-key header
Summary:
previous diffs added support for managing api keys in pixie.
this diff actually accepts the usage of api keys to access the cloud API.

Test Plan: unit test. launched cloud, made a cloud request with no auth token, but an API key in the headers instead. verified that the request was authorized. also tried with an invalid API key, and verified the key was rejected.

Reviewers: zasgar, philkuz, #engineering, vihang

Reviewed By: #engineering, vihang

Subscribers: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7006

GitOrigin-RevId: 1079cffa23079e533be44d4a2e1f0b64068c7895